update_triggers (trigger.c complete)
