+++ include/common_structs.h
@@ -271,11 +271,13 @@ typedef union {
     s32 flags;
 } TriggerFlags;
 
+typedef s32 (*TriggerHandlerFunc)(struct Trigger*);
+
 typedef struct Trigger {
     /* 0x00 */ TriggerFlags flags;
     /* 0x04 */ s32 params1;
     /* 0x08 */ s32 params2;
-    /* 0x0C */ UNK_FUN_PTR(functionHandler);
+    /* 0x0C */ TriggerHandlerFunc functionHandler;
     /* 0x10 */ EvtSource* scriptSource;
     /* 0x14 */ struct Evt* runningScript;
     /* 0x18 */ s32 priority;
@@ -1483,8 +1485,6 @@ typedef struct FontRasterSet {
     /* 0x02 */ char unk_02[10];
 } FontRasterSet; // size = 0x0C
 
-typedef void(*TriggerHandlerFunc)(struct Trigger*);
-
 typedef struct TriggerDefinition {
     /* 0x00 */ s32 flags;
     /* 0x04 */ s16 colliderIndex;
@@ -1512,7 +1512,7 @@ typedef struct CollisionStatus {
     /* 0x16 */ s16 touchingWallTrigger; /* 0/1 */
     /* 0x18 */ s16 bombetteExploded; /* 0 = yes, FFFF = no */
     /* 0x1A */ char unk_1A[2];
-    /* 0x1C */ f32 bombetteExplositionPos[3];
+    /* 0x1C */ Vec3f bombetteExplositionPos;
 } CollisionStatus; // size = 0x28
 
 typedef struct DecorationTable {
