+++ src/battle/partner/goombario.c
@@ -29,10 +29,10 @@ extern EvtSource N(idle_80239350);
 extern EvtSource N(nextTurn_80239A3C);
 extern EvtSource N(takeTurn_802396D8);
 
-void func_802380E4_6F11C4();
-void func_8023817C_6F125C();
-void func_80238BCC_6F1CAC();
-void func_80238EDC_6F1FBC();
+ApiStatus func_802380E4_6F11C4(Evt* script, s32 isInitialCall);
+ApiStatus func_8023817C_6F125C(Evt* script, s32 isInitialCall);
+ApiStatus func_80238BCC_6F1CAC(Evt* script, s32 isInitialCall);
+ApiStatus func_80238EDC_6F1FBC(Evt* script, s32 isInitialCall);
 
 ApiStatus N(func_80238000_6F10E0)(Evt* script, s32 isInitialCall) {
     BattleStatus* battleStatus = &gBattleStatus;
